requirements.txt
Django==4.2.4
djangorestframework==3.16.1
requests==2.32.5

